requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
langgraph==0.2.56
langchain-core==0.3.28
google-generativeai==0.8.3
httpx==0.27.2
tenacity==9.0.0
python-dotenv==1.0.1
pydantic==2.10.4
